+++ logic/include/emap/runconfiguration.h
@@ -67,6 +67,8 @@ public:
     fs::path scalings_path() const;
 
     const fs::path& data_root() const noexcept;
+    void set_data_root(const fs::path& root);
+
     const fs::path& output_path() const noexcept;
     const fs::path& spatial_pattern_exceptions() const noexcept;
     fs::path boundaries_vector_path() const noexcept;